:recycle: Refactor suggéré par Flo
